Use matchaddpos() to add BlingHilight highlight group
This happens to be faster, but it also fixes a bug with certain
patterns.  For example, let's say you search for the pattern "*".  Even
though this is technically illegal (the * pattern metacharacter needs to
follow an argument), Vim helpfully infers that you mean to search for
"*" on its own.   The problem is that when vim-bling does its thing, it
pads the current search pattern with a bunch of other stuff to make sure
only the match at the cursor position is found on the subsequent call to
matchadd(), and Vim is not so forgiving of the technically illegal
pattern here.  As a result, you get a bunch of error messages about
"misplaced *" or something like that, depending on which regexp engine
Vim is using, and the error repeats itself while vim-bling tries to
toggle the match on and off.  You can see this error in action here:

https://hoelz.ro/files/vim-bling-bug.gif

This patch uses Vim's own forgiving search logic to just find
the start and end positions of the match, and then passes them to
matchaddpos()